# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "FCT AUTHENTICATION BLOC",
        "description": "authentification api",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api",
            "description": "server"
        }
    ],
    "paths": {
...
        "/send-otp": {
            "post": {
                "tags": [
                    "OTP"
                ],
                "summary": "Allow to send otp ",
                "description": "You can send OTP by different channel like SMS, WHATSAPP and EMAIL",
                "operationId": "Send OTP",
                "requestBody": {
                    "description": "body request",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/SendOTP"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ResponseData"
                                }
                            },
                            "application/xml": {
                                "schema": {
                                    "$ref": "#/components/schemas/ResponseData"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "419": {
                        "description": "Expired session"
                    },
                    "404": {
                        "description": "Not found"
                    },
                    "500": {
                        "description": "Server Error"
                    }
                },
                "security": [
                    {
                        "JWT": []
                    }
                ]
            }
        },
        "/verify-otp": {
            "post": {
                "tags": [
                    "OTP"
                ],
                "summary": "Verify to vérification otp ",
                "description": "You can verify OTP send to channel like SMS, WHATSAPP and EMAIL",
                "operationId": "Verify OTP",
                "requestBody": {
                    "description": "body request",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/VerifyOTP"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ResponseData"
                                }
                            },
                            "application/xml": {
                                "schema": {
                                    "$ref": "#/components/schemas/ResponseData"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "419": {
                        "description": "Expired session"
                    },
                    "404": {
                        "description": "Not found"
                    },
                    "500": {
                        "description": "Server Error"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "ResponseData": {
                "title": " Output",
                "description": "Output model",
                "properties": {
                    "message": {
                        "type": "string"
                    },
                    "status": {
                        "type": "boolean"
                    },
                    "data": {
                        "type": "string"
                    },
                    "warning": {
                        "type": "string"
                    },
                    "error_list": {
                        "type": "array",
                        "items": {
                            "properties": {
                                "id": {
                                    "description": "",
                                    "type": "string"
                                },
                                "response": {
                                    "description": "",
                                    "type": "string"
                                }
                            },
                            "type": "object"
                        }
                    }
                },
                "type": "object"
            },
...
            "SendOTP": {
                "title": "SendOTP class",
                "description": "SendOTP class",
                "properties": {
                    "phone": {
                        "type": "integer"
                    },
                    "email": {
                        "type": "string"
                    },
                    "canal": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        }
                    }
                },
                "type": "object"
            },
...
            "VerifyOTP": {
                "title": "VerifyOTP class",
                "description": "VerifyOTP class",
                "properties": {
                    "phone": {
                        "type": "integer"
                    },
                    "email": {
                        "type": "string"
                    },
                    "verification_code": {
                        "type": "string"
                    },
                    "for_login": {
                        "type": "integer"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "JWT": {
                "type": "apiKey",
                "name": "Authorization",
                "in": "header",
                "scheme": "bearer"
            }
        }
    },
...
}